+++ tgbotapi.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/utils/SimpleFilter.kt
@@ -1,8 +1,8 @@
 package dev.inmo.tgbotapi.extensions.behaviour_builder.utils
 
-typealias SimpleFilter<T> = suspend (T) -> Boolean
-
-inline fun <T> SimpleFilter(noinline block: SimpleFilter<T>) = block
+fun interface SimpleFilter<in T> {
+    suspend operator fun invoke(o: T): Boolean
+}
 
 /**
  * @return [SimpleFilter] which will return true in case when all the items in incoming data passed [this] filter
@@ -28,20 +28,20 @@ fun <T> SimpleFilter<T>.listNone() = SimpleFilter<Iterable<T>> {
 /**
  * Makes an AND (&&) operation between [this] and [other]
  */
-operator fun <T> SimpleFilter<T>.times(other: SimpleFilter<T>): SimpleFilter<T> = {
+operator fun <T> SimpleFilter<T>.times(other: SimpleFilter<T>) = SimpleFilter<T> {
     this(it) && other(it)
 }
 
 /**
  * Makes an OR (||) operation between [this] and [other]
  */
-operator fun <T> SimpleFilter<T>.plus(other: SimpleFilter<T>): SimpleFilter<T> = {
+operator fun <T> SimpleFilter<T>.plus(other: SimpleFilter<T>) = SimpleFilter<T> {
     this(it) || other(it)
 }
 
 /**
  * Reverse results of [this]
  */
-operator fun <T> SimpleFilter<T>.not(): SimpleFilter<T> = {
+operator fun <T> SimpleFilter<T>.not() = SimpleFilter<T> {
     !this(it)
 }